AI sentiment analysis
In these files is the code for an university assignment about AI sentiment analysis. These are made with python and some other python libraries like pytorch or numpy. There are 3 algorithms here, Naive Bayes, Logistic Regression and a multilayered bi-directional RNN with LSTM cells. All are trained on the IMDB database which has about 25000 reviews on movies. Their purpose is to figure out if a review is positive or negative.